Desarrollo de  registro de votos

+++ src/main/java/com/hakaton/repository/MunicipioRepository.java
@@ -3,5 +3,5 @@ package com.hakaton.repository;
 import com.hakaton.model.Municipio;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MunicipioRepository extends JpaRepository<Municipio, Long> {
+public interface MunicipioRepository extends JpaRepository<Municipio, Integer> {
 }
